+++ spec/rubocop/cop/lint/literal_in_condition_spec.rb
@@ -115,4 +115,40 @@ describe Rubocop::Cop::Lint::LiteralInCondition do
       expect(cop.offenses).to be_empty
     end
   end
+
+  it 'accepts array literal in case, if it has non-literal elements' do
+    inspect_source(cop,
+                   ['case [1, 2, x]',
+                    'when [1, 2, 5] then top',
+                    'end'
+                   ])
+    expect(cop.offenses).to be_empty
+  end
+
+  it 'accepts array literal in case, if it has non-literal elements' do
+    inspect_source(cop,
+                   ['case [1, 2, [x, 1]]',
+                    'when [1, 2, 5] then top',
+                    'end'
+                   ])
+    expect(cop.offenses).to be_empty
+  end
+
+  it 'registers an offense for case with a primitive array condition' do
+    inspect_source(cop,
+                   ['case [1, 2, [3, 4]]',
+                    'when [1, 2, 5] then top',
+                    'end'
+                   ])
+    expect(cop.offenses.size).to eq(1)
+  end
+
+  it 'accepts dstr literal in case' do
+    inspect_source(cop,
+                   ['case "#{x}"',
+                    'when [1, 2, 5] then top',
+                    'end'
+                   ])
+    expect(cop.offenses).to be_empty
+  end
 end